Create new UI package for view fragments
