+++ app/src/main/java/com/github/dedis/student20_pop/ui/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop;
+package com.github.dedis.student20_pop.ui;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -8,6 +8,9 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+
+import com.github.dedis.student20_pop.R;
+import com.github.dedis.student20_pop.ui.ConnectFragment;
 
 /**
  * Fragment used to display the Connect UI
