+++ app/src/main/java/com/github/dedis/student20_pop/ui/LaunchFragment.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop;
+package com.github.dedis.student20_pop.ui;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,16 +9,18 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-/**
- * Fragment used to display the Connect UI
-**/
-public final class ConnectFragment extends Fragment {
+import com.github.dedis.student20_pop.R;
 
-    public static final String TAG = ConnectFragment.class.getSimpleName();
+/**
+ * Fragment used to display the Launch UI
+**/
+public final class LaunchFragment extends Fragment {
+
+    public static final String TAG = LaunchFragment.class.getSimpleName();
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_connect, container, false);
+        return inflater.inflate(R.layout.fragment_launch, container, false);
     }
 }
