+++ app/src/main/java/com/github/dedis/student20_pop/ui/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Classes used to represent the different UIs
+ * <p>
+ * The classes inside this package are fragments.
+ * </p>
+ **/
+package com.github.dedis.student20_pop.ui;
